+++ docs/RELEASE-NOTES.md
@@ -1,5 +1,19 @@
 # Release Notes
 
+## 1.1.0 - 11 December 2020
+
+Breaking changes:
+- (None)
+
+New features:
+- Added `Base32Encoder`
+- Added `Base32CharacterSet`
+- Added `Base32CharExtensions.IsBase32`
+- Added `Base32StringExtensions.IsBase32`
+
+Bug fixes / internal changes:
+- (None)
+
 ## 1.0.1 - 06 October 2020
 
 Breaking changes: